Json working correctly

+++ src/main/java/Player.java
@@ -1,4 +1,3 @@
-import java.lang.reflect.Array;
 import java.util.*;
 
 /*////////////////////////////////////////////////////////////////////////////////
@@ -12,7 +11,7 @@ public class Player {
     private Army placeholder;
     private List<String> territories;
     private HashMap<String, Stack<Card>> Hand;
-    private float wallet = 0;
+    private Double wallet = 0.0;
     private int Undos = 0;
 
     // add undos
@@ -22,10 +21,10 @@ public class Player {
     public int getUndos() { return Undos; }
 
     // add to the wallet
-    public void addMoney(float add) { wallet = wallet + add; }
+    public void addMoney(double add) { wallet = wallet + add; }
 
     // get wallet contents
-    public float getWallet() { return wallet; }
+    public Double getWallet() { return wallet; }
 
     // get player id
     public int getId() {
